Fix some warnings

+++ src/IOST/Signer.cpp
@@ -33,12 +33,12 @@ class IOSTEncoder {
     }
 
     void WriteString(std::string s) {
-        WriteInt32(s.size());
+        WriteInt32(static_cast<uint32_t>(s.size()));
         buffer << s;
     }
 
     void WriteStringSlice(const std::vector<std::string> v) {
-        WriteInt32(v.size());
+        WriteInt32(static_cast<uint32_t>(v.size()));
         for (std::string s : v) {
             WriteString(s);
         }
